feat: use external Header component
